requirements.txt
jsonschema==2.3.0
six==1.10.0
requests>=2.11
responses==0.5.1
boto3>=1.4.0
moto==0.4.25
Pillow>=3.3.1
numpy>=1.11.1
nose2==0.6.5
cloud-volume